refactor(components) move all individual js files into their own folder
