+++ components/DateDisplay/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text } from 'react-native';
-import { toRelativeDate } from '../services/date';
+import { toRelativeDate } from '../../services/date';
 
 export default function DateDisplay(props) {
     const { date } = props;
